pyproject.toml
[project]
name = "goit-pythonweb-hw-12"
version = "0.1.0"
description = "Contacts API HW12"
authors = [
    {name = "vladshein",email = "vlshein1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "alembic (>=1.16.5,<2.0.0)",
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "libgravatar (>=1.0.4,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "fastapi-mail (>=1.5.0,<2.0.0)",
    "cloudinary (>=1.44.1,<2.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "redis-lru (>=0.1.2,<0.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
sphinx = "^8.2.3"

